requirement.txt
Django==1.11.5
django-autoslug==1.9.3
django-cities-light==3.3.0
olefile==0.44
Pillow==4.2.1
progressbar2==3.34.3
psycopg2==2.7.3.1
python-utils==2.2.0
pytz==2017.2
six==1.11.0
Unidecode==0.4.21
